[fix] types

+++ include/types.h
@@ -18,7 +18,7 @@ typedef uint64_t bitflag64_t;
 typedef uint8_t  u8;
 typedef uint16_t u16;
 typedef uint32_t u32;
-typedef uint64_t u64;
+typedef unsigned long long u64;
 typedef int8_t   i8;
 typedef int16_t  i16;
 typedef int32_t  i32;
